rename withJWTAuthHandler wrapper files
