+++ apps/vehicles/src/lib/bff/with-jwt-auth-handler.ts
